Added individual pictures for legend
These pictures are about the height of an uppercase 'M', so they can fit within prose without looking too weird. It's worth noting that, for the "N9" symbol, I used text, and it does not change sizes with the `scale` option. That's why it's `\tiny` in order to fit within the box. If you make the symbols larger for your legend, you might want to change that text size (which is the default size in the actual chart).